Rename logic modules more descriptively
