+++ src/game/mod.rs
@@ -3,13 +3,13 @@
 pub mod animation;
 pub mod components;
 pub mod drawing;
-mod ecs_free_logic;
 mod game_sfx;
 mod hints;
 pub mod history;
 mod inputs;
 pub mod level;
 pub mod logic;
+pub mod logic_relay;
 pub mod spawn;
 mod synchronization;
 #[cfg(test)]
@@ -18,10 +18,10 @@ pub mod test;
 pub mod prelude {
 	#[allow(unused_imports)]
 	pub use super::{
-		ecs_free_logic::{GameState, LevelCompletionConditions},
 		history::{MoveHistory, UndoMove},
 		level::LevelData,
-		logic::{CycleTurningDirection, GameLayoutChanged, IsLevelCompleted, RotateCycle},
+		logic::{GameState, LevelCompletionConditions},
+		logic_relay::{CycleTurningDirection, GameLayoutChanged, IsLevelCompleted, RotateCycle},
 		spawn::{
 			ActiveLevel, ActiveLevelData, EnterLevel, LevelInitialization, LevelInitializationSet,
 		},
@@ -35,7 +35,7 @@ use bevy::prelude::*;
 pub(super) fn plugin(app: &mut App) {
 	app.add_plugins((
 		spawn::plugin,
-		logic::plugin,
+		logic_relay::plugin,
 		animation::plugin,
 		drawing::plugin,
 		inputs::plugin,
